+++ test/Matchers/SpatialTest.php
@@ -126,14 +126,14 @@ class SpatialTest extends AbstractMatchTest
         );
     }
 
-    protected function getBaseGuessCount($token): int
+    protected function getBaseGuessCount($token): float
     {
         // KEYBOARD_STARTING_POSITIONS * KEYBOARD_AVERAGE_DEGREE * (length - 1)
         // - 1 term because: not counting spatial patterns of length 1
         // eg for length==6, multiplier is 5 for needing to try len2,len3,..,len6
-        return (int)(SpatialMatch::KEYBOARD_STARTING_POSITION
+        return SpatialMatch::KEYBOARD_STARTING_POSITION
             * SpatialMatch::KEYBOARD_AVERAGE_DEGREES
-            * (strlen($token) - 1));
+            * (strlen($token) - 1);
     }
 
     public function testGuessesBasic()
